pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "releasepilot"
version = "1.0.0"
description = "Production-grade release notes and release communication tool"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [{name = "ReleasePilot Contributors"}]
keywords = ["release-notes", "changelog", "git", "automation", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Version Control :: Git",
    "Typing :: Typed",
]
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "questionary>=2.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]
export = [
    "reportlab>=4.0",
    "python-docx>=1.0",
]
translate = [
    "deep-translator>=1.11",
]
all = [
    "releasepilot[dev,export,translate]",
]

[project.scripts]
releasepilot = "releasepilot.cli.app:main"

[project.urls]
Homepage = "https://polprog.pl/apps/ReleasePilot"
Repository = "https://github.com/polprog-tech/ReleasePilot"
Issues = "https://github.com/polprog-tech/ReleasePilot/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
releasepilot = ["dashboard/templates/*.j2"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["N801", "SIM117"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
source = ["releasepilot"]
omit = ["tests/*"]
